requirements.txt
gensim
numpy
Flask
python-decouple
requests
fasttext
torch
torchvision
Pillow
tensorflow
Werkzeug
spicy